+++ client/packages/host/src/components/Navigation/ReplenishmentNav.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React from 'react';
 
 import {
   SuppliersIcon,
@@ -8,15 +8,21 @@ import {
   RouteBuilder,
   AppNavLink,
   AppNavSection,
+  UserStoreNodeFragment,
 } from '@openmsupply-client/common';
 import { AppRoute } from '@openmsupply-client/config';
 import { useNestedNav } from './useNestedNav';
 
-export const ReplenishmentNav: FC = () => {
+export const ReplenishmentNav = ({
+  store,
+}: {
+  store?: UserStoreNodeFragment;
+}) => {
   const { isActive } = useNestedNav(
     RouteBuilder.create(AppRoute.Replenishment).addWildCard().build()
   );
   const t = useTranslation('app');
+  const rnrVisible = store?.preferences.omProgramModule;
 
   return (
     <AppNavSection isActive={isActive} to={AppRoute.Replenishment}>
@@ -49,6 +55,14 @@ export const ReplenishmentNav: FC = () => {
               .addPart(AppRoute.OutboundReturn)
               .build()}
             text={t('outbound-returns')}
+          />
+          <AppNavLink
+            visible={rnrVisible}
+            end
+            to={RouteBuilder.create(AppRoute.Replenishment)
+              .addPart(AppRoute.RnRForms)
+              .build()}
+            text={t('r-and-r-forms')}
           />
           <AppNavLink
             end
