+++ client/packages/programs/src/Service/Service.tsx
@@ -3,7 +3,6 @@ import { RouteBuilder, Routes, Route } from '@openmsupply-client/common';
 import { AppRoute } from '@openmsupply-client/config';
 import { ImmunisationProgramListView } from '../ImmunisationProgramListView';
 import { ImmunisationProgramDetailView } from '../ImmunisationProgramDetailView';
-import { RnRFormDetailView, RnRFormListView } from '../RnRForms';
 
 export const ProgramService: FC = () => {
   const immunisationProgramsRoute = RouteBuilder.create(
@@ -13,12 +12,6 @@ export const ProgramService: FC = () => {
   const immunisationProgramRoute = RouteBuilder.create(
     AppRoute.ImmunisationPrograms
   )
-    .addPart(':id')
-    .build();
-
-  const rnrFormsRoute = RouteBuilder.create(AppRoute.RnRForms).build();
-
-  const rnrFormRoute = RouteBuilder.create(AppRoute.RnRForms)
     .addPart(':id')
     .build();
 
@@ -32,8 +25,6 @@ export const ProgramService: FC = () => {
         path={immunisationProgramRoute}
         element={<ImmunisationProgramDetailView />}
       />
-      <Route path={rnrFormsRoute} element={<RnRFormListView />} />
-      <Route path={rnrFormRoute} element={<RnRFormDetailView />} />
     </Routes>
   );
 };
