+++ client/packages/programs/src/api/hooks/index.ts
@@ -17,11 +17,6 @@ export * from './useImmunisationProgram';
 export * from './useImmunisationProgramList';
 export * from './useVaccineCourse';
 export * from './useVaccineCourseList';
-export * from './useRnRForm';
-export * from './useRnRFormList';
-export * from './useProgramList';
-export * from './useSchedulesAndPeriods';
-export * from './useCreateRnRForm';
 
 export {
   ProgramEnrolmentRowFragmentWithId,
