+++ client/packages/programs/src/api/hooks/keys.ts
@@ -2,5 +2,3 @@ export const IMMUNISATION_PROGRAM = 'immunisation-program';
 export const PROGRAM = 'program';
 export const LIST = 'list';
 export const VACCINE = 'course';
-export const RNR_FORM = 'rnr-form';
-export const SCHEDULE = 'schedule';
