+++ client/packages/requisitions/src/RnRForms/ListView/ListView.tsx
@@ -15,10 +15,10 @@ import {
 } from '@openmsupply-client/common';
 import { Toolbar } from './Toolbar';
 import { AppBarButtons } from './AppBarButtons';
-import { useRnRFormList } from '../../api';
-import { RnRFormFragment } from '../../api/operations.generated';
+import { useRnRFormList } from '../api';
+import { RnRFormFragment } from '../api/operations.generated';
 import { RnRFormCreateModal } from './RnRFormCreateModal';
-import { getStatusTranslator, isRnRFormDisabled } from '../../utils';
+import { getStatusTranslator, isRnRFormDisabled } from '../utils';
 
 const RnRFormListComponent = () => {
   const {
