+++ client/packages/requisitions/src/RnRForms/api/hooks/keys.ts
@@ -1,0 +1,4 @@
+export const RNR_FORM = 'rnr-form';
+export const SCHEDULE = 'schedule';
+export const PROGRAM = 'program';
+export const LIST = 'list';
