+++ client/packages/requisitions/src/RnRForms/api/useRnRGraphQL.ts
@@ -1,0 +1,15 @@
+import {
+  useAuthContext,
+  useGql,
+  useQueryClient,
+} from '@openmsupply-client/common';
+import { getSdk } from './operations.generated';
+
+export const useRnRGraphQL = () => {
+  const { client } = useGql();
+  const queryClient = useQueryClient();
+  const { storeId } = useAuthContext();
+  const api = getSdk(client);
+
+  return { api, queryClient, storeId };
+};
